# specs/001-finance-hub/contracts/api.yaml
openapi: 3.1.0
info:
  title: Finance Hub API
  version: 1.0.0
  description: Self-hosted financial management API
  contact:
    name: Finance Hub
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        deviceName:
          type: string
    LoginResponse:
      type: object
      properties:
        requiresMfa:
          type: boolean
        mfaMethods:
          type: array
          items:
            type: object
            properties:
              id:
                $ref: '#/components/schemas/UUID'
              type:
                type: string
                enum: [totp, webauthn]
              name:
                type: string
        tempToken:
          type: string
          description: Temporary token for MFA step
    MfaVerifyRequest:
      type: object
      required: [tempToken, mfaId]
      properties:
        tempToken:
          type: string
        mfaId:
          $ref: '#/components/schemas/UUID'
        code:
          type: string
          description: TOTP code (for totp type)
        credential:
          type: object
          description: WebAuthn credential (for webauthn type)
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        tokenType:
          type: string
          default: Bearer

    # User
...
    CreateAccountRequest:
      type: object
      required: [name, type, currency]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, cash, credit, investment, loan]
        currency:
          type: string
        initialBalance:
          type: number
          default: 0
        institution:
          type: string
        color:
          type: string
        icon:
          type: string

    # Transaction
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Verify MFA code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerifyRequest'
      responses:
        '200':
          description: MFA verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid MFA code
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
  /auth/unlock:
    post:
      tags: [Auth]
      summary: Unlock session with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
      responses:
        '204':
          description: Session unlocked

  # User
...
  /workspaces/{workspaceId}/accounts:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'

    get:
      tags: [Accounts]
      summary: List accounts
      responses:
        '200':
          description: Account list

    post:
      tags: [Accounts]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created

  # Transactions
...